Error on gdt table, on initialization error

+++ src/common.hh
@@ -3,13 +3,13 @@
 typedef unsigned char byte;
 typedef char sbyte;
 
-typedef unsigned char  uint8;
-typedef unsigned short uint16;
-typedef unsigned int   uint32;
-typedef unsigned long  uint64;
-typedef char           int8;
-typedef short          int16;
-typedef int            int32;
-typedef long           int64;
+typedef unsigned char       uint8;
+typedef unsigned short      uint16;
+typedef unsigned int        uint32;
+typedef unsigned long long  uint64;
+typedef char                int8;
+typedef short               int16;
+typedef int                 int32;
+typedef long long           int64;
 
 #define BIT(x) (1 << (x))
